pyproject.toml
[project]
name = "redkeds"
version = "0.1.0"
description = "Social designer web site"
readme = "README.md"
authors = [
    { name = "Aliaksei Yafremau", email = "aleksey.yafremau@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "dishka>=1.6.0",
    "fastapi>=0.115.13",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[project.scripts]
run = "redkeds.main.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
test = [
    "coverage>=7.10.1",
    "faker>=37.5.3",
    "pytest-asyncio>=1.0.0",
]
lint = [
    "ruff>=0.11.13",
]
type-check = [
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "dishka>=1.6.0",
    "fastapi>=0.115.13",
    "mypy>=1.16.1",
    "pyjwt>=2.10.1",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
exclude = [
	# Игнорировать папку "tests" и все её подпапки
	"^tests/",
    # Игнорировать папку "scripts" и все её подпапки
    "^scripts/",
    # Игнорировать папки миграций alembic
    "src/redkeds/adapters/migrations/versions"
]

[tool.ruff]
exclude = [
    # Игнорировать папки миграций alembic
    "src/redkeds/adapters/migrations/versions"
]

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "D100", # Не нужно писать docstring для модулей
    "D104", # Не нужно писать docstring для пакетов
    "D107", # Не нужно писать docstring для __init__()
    "D213", # Не нужно начинать docstring с второй строки
    "D203", # Не нужно писать docstring с первой строки
    "S101", # Возможность использования assert
    "EM101", # Возможность использования строки при вызове ошибки
    "EM102", # Возможность использования f-строки при вызове ошибки
    "PLR2004", # Возможность использования "magic" значений
    "RUF001", # Игнорирование неоднозначных символов в строке
    "RUF002", # Игнорирование неоднозначных символов в docstring
    "RUF003", # Игнорирование неоднозначных символов в коментарии
    "COM812", # Не нужно писать запятую в конце списка
    "TRY003", # Возможность писать сообщение в ошибке
    "S106", # Возможность hardcode пароль для тестовых реализаций
    "B904", # Возможность вызова exception без 'from'
    "ANN202", # Не нужно писать docstring для не публичного API
    "ANN401", # Возможность аннотировать с помощью Any
    "D105", # Не нужно писать docstring для магических методов
    "PT014", # Возможность дупликации при параметризации тестов
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D103", "PLR0913", "S311", "SLF001"] 

[tool.alembic]
script_location = "%(here)s/src/redkeds/adapters/migrations"

# дополнительные пути для sys.path
prepend_sys_path = [
    "."
]

# спецификация местоположения версий (по умолчанию: <script_location>/versions)
version_locations = [
    "%(here)s/src/redkeds/adapters/migrations/versions"
]

# кодировка вывода, используемая при записи файлов миграций из script.py.mako
output_encoding = "utf-8"
